% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidatesheets.R
\name{consolidatesheets}
\alias{consolidatesheets}
\title{Consolidate sheets of an Excel workbook into a single data frame}
\usage{
consolidatesheets(path, verbose = FALSE)
}
\arguments{
\item{path}{file name of the source Excel workbook}

\item{verbose}{when \code{TRUE} display \code{str} of returned data frame}
}
\value{
a single data frame containing all sheets
}
\description{
\code{consolidatesheets} is useful for combining multiple sheets of an
Excel workbook into a single data frame.  It treats all columns as text
and assumes that the sheets all have the same structure.
}
