% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateworkbooks.R
\name{consolidateworkbooks}
\alias{consolidateworkbooks}
\title{Consolidate multiple Excel workbooks into a single workbook file}
\usage{
consolidateworkbooks(workbookfilenames, targetfilename, verbose = FALSE)
}
\arguments{
\item{workbookfilenames}{a list of file names to be consolidated}

\item{targetfilename}{the new file to be written as a single page workbook}

\item{verbose}{when \code{TRUE} display processing information}
}
\value{
a single data frame containing all sheets
}
\description{
\code{consolidateworkbooks} takes a list of Excel filenames, reads and
consolidates the files, writes a new Excel workbook, which
contains all of the source workbooks and their sheets in a single
sheet.
}
