% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolxl.R
\name{consolxl}
\alias{consolxl}
\title{Consolidate multiple Excel workbooks into a single workbook file}
\usage{
consolxl(verbose = FALSE)
}
\arguments{
\item{verbose}{when \code{TRUE} display processing information}
}
\value{
a single data frame containing all sheets
}
\description{
\code{consolxl} combine multiple Excel workbooks and their contained
sheets into a single Excel file. This function assumes all selected
workbooks and their sheets have the same structure.  The user is
prompted for source files and target file name.
}
