% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopquietly.R
\name{stopquietly}
\alias{stopquietly}
\title{Stop processing quietly (without error)}
\usage{
stopquietly(message, ...)
}
\arguments{
\item{message}{the message to display to the user when stopping}

\item{...}{further arguments for \code{print} function}
}
\description{
Stop the processing with a message.  Usually executed in the context
of a user making a cancel decision during interaction.
}
