% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateworkbooks.R
\name{consolidateworkbooks}
\alias{consolidateworkbooks}
\title{Consolidate multiple Excel workbooks into a single workbook file}
\usage{
consolidateworkbooks(verbose = FALSE)
}
\arguments{
\item{verbose}{when \code{TRUE} display processing information}
}
\value{
a single data frame containing all sheets
}
\description{
\code{consolidateworkbooks} is useful for combining multiple
Excel workbooks and their contained sheets into a single Excel file.
This function was coded with the assumption all of the workbooks
and their contained sheets have the same structure.  The user is
prompted for source files and target file name.
}
